created two distinct files for utilities
