+++ tests/test_remotepath.py
@@ -8,7 +8,7 @@ from streamflow.core.data import FileType
 from streamflow.core.deployment import Connector, Location
 from streamflow.data import remotepath
 from streamflow.deployment.utils import get_path_processor
-from tests.conftest import get_location
+from tests.utils.deployment import get_location
 
 
 @pytest_asyncio.fixture(scope="module")
