+++ tests/test_transfer.py
@@ -12,7 +12,7 @@ from streamflow.core.deployment import Connector, Location
 from streamflow.data import remotepath
 from streamflow.deployment.connector import LocalConnector
 from streamflow.deployment.utils import get_path_processor
-from tests.conftest import get_location
+from tests.utils.deployment import get_location
 
 
 @pytest_asyncio.fixture(scope="module")
